moved grid cell component to new file
